Add composer.json file
This allows this repo to used as a composer project (see https://getcomposer.org/doc/04-schema.md#type). This means that you can do something like this to start a new mini-extension:

```
composer create-project prospress/mini-extension-boilerplate
```

This is a better option than the current process described in the wiki of cloning, changing the remote, and pushing to a new repo.

We do still need to publish this packagist.org for that to work fully.